app/requirements.txt
fastmcp==2.12.4
python-dotenv>=1.0.0